requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
apache-airflow == 2.8.0
psycopg2-binary